Create SVD.py
Created a class to cast a given matrix into its SVD form. Still to-do is sort the different arrays accordingly and implement the finite sum afterwards. The test method in the class can easily be adapted to the latter.

Changed the name of the python testing directory to testing (-Yovany)